chore: Fix up owlbot scripts for vision gapics
